fastapi_assignment_2/requirements.txt

fastapi~=0.95.1
pydantic~=1.10.4
pymongo~=4.3.3
uvicorn~=0.22.0

python-dotenv~=1.0.0
pyyaml~=6.0